# swagger.yaml
basePath: /
consumes:
    - application/json
definitions:
    GenericError:
        description: GenericError is a generic error message returned by a server
        properties:
            message:
                type: string
                x-go-name: Message
        type: object
        x-go-package: github.com/ruan-molinari/go-microservices/handlers
...
info:
    description: '# Documentation for Product API'
    title: of Product API
    version: 0.1.0
paths:
...
    /products/{id}:
        delete:
            description: Delete a product from the database
            operationId: deleteProduct
            parameters:
                - description: The id of the product to delete from database
                  format: int64
                  in: path
                  name: id
                  required: true
                  type: integer
                  x-go-name: ID
            responses:
                "201":
                    $ref: '#/responses/noContent'
                "404":
                    $ref: '#/responses/errorResponse'
                "501":
                    $ref: '#/responses/errorResponse'
            tags:
                - products
produces:
    - application/json
responses:
    errorResponse:
        description: Generic error message returned as string
        schema:
            $ref: '#/definitions/GenericError'
...
schemes:
    - http
swagger: "2.0"